requirements.txt
# Core dependencies
accelerate==1.12.0
bitsandbytes==0.49.1
einops==0.7.0
huggingface-hub==0.24.7
numpy==2.1.3
sentencepiece==0.2.0
sphn==0.1.12
torch==2.4.1+cu121

# GUI dependencies
gradio>=4.0.0
pyyaml>=6.0

# Local package
-e .
